增加订单日志链接

+++ Payment/WX/Pay.php
@@ -34,7 +34,7 @@ $input -> SetBody("微信支付金额:".$money);
 $input -> SetAttach("微信支付");
 //$input -> SetOut_trade_no(WxPayConfig::MCHID . date("YmdHis"));
 $input -> SetOut_trade_no($orderid);
-$input -> SetTotal_fee(intval($money * 100));
+$input -> SetTotal_fee(floatval($money * 100));
 //$input -> SetTotal_fee("1");
 $input -> SetTime_start(date("YmdHis"));
 $input -> SetTime_expire(date("YmdHis", time() + 600));
